rename benchmarks & separate generate_parameters
