+++ manta-benchmark/Cargo.toml
@@ -20,7 +20,6 @@ all-features = true
 rustdoc-args = ["--cfg", "doc_cfg"]
 
 [dependencies]
-anyhow = { version = "1.0.55", default-features = false }
 manta-accounting = { path = "../manta-accounting", default-features = false, features = ["test"] }
 manta-crypto = { path = "../manta-crypto", default-features = false, features = ["test"] }
 manta-pay = { path = "../manta-pay", default-features = false, features = ["groth16", "test"] }
@@ -30,13 +29,13 @@ rand_chacha = { version = "0.3.1", default-features = false }
 criterion = { version = "0.3.4", default-features = false }
 
 [[bench]]
-name = "bench_mint"
+name = "mint"
 harness = false
 
 [[bench]]
-name = "bench_private_transfer"
+name = "private_transfer"
 harness = false
 
 [[bench]]
-name = "bench_reclaim"
+name = "reclaim"
 harness = false
