+++ manta-benchmark/src/lib.rs
@@ -14,5 +14,4 @@
 // You should have received a copy of the GNU General Public License
 // along with manta-rs.  If not, see <http://www.gnu.org/licenses/>.
 
-pub mod parameters;
 pub mod payment;
