+++ manta-pay/src/crypto/mod.rs
@@ -21,3 +21,4 @@ pub mod ecc;
 pub mod encryption;
 pub mod hash;
 pub mod key;
+pub mod parameters;
